requirements.txt
Flask==2.2.2
PyPDF2==3.0.0
requests==2.28.1
python-dotenv==1.0.0
openai==0.27.0
